pyproject.toml
[project]
name = "car_dealer"
version = "0.1.0"
description = " Car dealer platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
    "mock>=5.1.0",
    "pytest-cov>=5.0.0",
    "pytest-django>=4.9.0",
    "pytest>=8.3.3",
    "ruff>=0.6.9",
]

[tool.uv]
dev-dependencies = []
